airbnb_reg/requirements.txt
appnope==0.1.2
asttokens==2.0.5
backcall==0.2.0
black==22.1.0
blis==0.7.5
catalogue==2.0.6
charset-normalizer==2.0.11
click==8.0.3
cycler==0.11.0
cymem==2.0.6
decorator==5.1.1
executing==0.8.2
fastai==2.5.3
fastcore==1.3.27
fastdownload==0.0.5
fastprogress==1.0.0
fonttools==4.29.0
idna==3.3
inplace-abn==1.1.0
ipython==8.0.1
jedi==0.18.1
jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
langcodes==3.3.0
markupsafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
murmurhash==1.0.6
mypy-extensions==0.4.3
numpy==1.22.1
packaging==21.3
pandas==1.4.0
parso==0.8.3
pathspec==0.9.0
pathy==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
pillow==9.0.0
platformdirs==2.4.1
preshed==3.0.6
prompt-toolkit==3.0.26
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.8.2
pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
pyyaml==6.0
requests==2.27.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
smart-open==5.2.1
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.2
stack-data==0.1.4
thinc==8.0.13
threadpoolctl==3.0.0
tomli==2.0.0
torch==1.10.2
torchvision==0.11.3
tqdm==4.62.3
traitlets==5.1.1
typer==0.4.0
typing-extensions==4.0.1
urllib3==1.26.8
wasabi==0.9.0
wcwidth==0.2.5
monk-cpu